backend/requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
multidict==6.4.3
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-dotenv==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.46.0
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.19.0

